refactor(errors): move efbig
